requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
python-docx>=1.1.0
langfuse>=2.20.0
ollama>=0.2.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.20.0
pdfplumber>=0.10.0
Pillow>=10.2.0
httpx>=0.27.0
